moved bottom bar to the main page
